Implementation of IPinIP tunneling. Reorganize code to improve the utilization of dp meta data.
This commit mainly concerns the implementation of IPvXinIPv6 tunneling in non-/offloading mode.
The integration of existing GENEVE impl is also included by providing a starting parameter --tun_opt
The lookup nodes and its neighboring nodes are modified to improve the orgnization of dp meta data fields to make it more self-explanatory.