Initial/simple implementation of WCMP on hypervisor.
This commit is initial/simple implementation of WCMP on hypervisor.
The idea of WCMP on hypervisor is that traffic flows coming out of a VM will be forwarded to different pf ports, depending on the hashing value of this flow.
It is achieved by using an additional parameter, --wcmp-frac=[0.0-1.0], when starting dp_service. This value indicates that what is the percentage of the traffic that need to be forwarded through this VF's owner PF.
When this value is set as 0.5, we have ECMP, in which, traffic workload is meant to be equally shared between PFs (in theory).
Due to the constraint of Mellanox nic/driver, in the offloading mode, the traffic flows that are determined to forwarded to the peer pf port, go through the software path.
With 2 core configuration for dp_service, there is a degredation of max throughput.