Reworked rte flow funcs to have a clean flow building. New dpdk init funcs and tx encap-decap funcs.
This commit contains a proposed way to construct desired rte flow rules.
The idea is to provide wrappers for rte_flow matching patterns and actions, so that it is possible to construct them in each individual rte flow rule with one function call.
In this commit, dpdk init flow rule and encap/decap rules in tx node are updated to use these wrappers.
Due to the constraints on the conn track for icmp packets, aging actions are disabled for now to allow icmp traffic to be offloaded.