This commit contains init implementation of failure handling (broken link).
This commit contains the init implementation of dynamic selection of outgoing pf ports when a link is detected as down.
It utilizes the link status interrupts to change the stored link status, and this information is considered in making routing decisions.
The data structure used to store info of grpc, monitoring, periodic queues, is rewritten to correct the misuse of graph nodes' ctx field.
The limitation is that failure handling happends in the software path, hence, for non-offloading mode, it works smoothly.
But migrating installed flow rules for tunneling or LB is not part of this implementation.