Rework the NAT related Get and Set functions
Local NAT operates with Interface ID as the key so no need
to operate with NAT VIP in the parameters when doing get and delete
operations. Also Getters for local and neighbour NAT needs
to be seperated to have a clean design and to make the life
on controller side easier.

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>