Identify whether the prepare.sh script on host or on bluefield runs
The Bluefield card is visible to the host x86 system where it is
connected and the prepare script also runs on the host system.
(not the dpservice but just the prepare script)

Therefore the script needs to be able to differentiate between
the cases running on the bluefield directly and on the host where
the bluefield is connected.

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>