Place iface_id as the first element in the struct so the alignment is correct on ARM
On ARM architecture iface_id in "struct dpgrpc_vip" seems to start on
odd memory address as it is not the first element of the struct where
it is defined and the element before it, is packed.

This breaks the ifaced_id hash table lookup in create vip grpc call. So reorder
the variables.

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>