Switch from /tmp to /run to DP_CONF default path
On modern debian-based systems files may be cleared from /tmp after some
time. The recommended location for files which should be kept as long as
the system is running is the /run directory.

This commit changes the relevant references to load the config file from
/run/dpservice/dpservice.conf by default instead of /tmp/dp_service.conf.
The old behavior can be restored by setting DP_CONF to
/tmp/dp_service.conf.

The hack/prepare.sh was updated to respect the DP_CONF environemnt
variable as well.

Resolves: #717